small fixes for mingw/win32 compilation
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@2237 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/kvirc/ui/kvi_channel.cpp
@@ -1134,7 +1134,7 @@ void KviChannel::ownMessage(const QString &buffer)
 				//if szTmp.length() == 0 we already have break'ed out from here,
 				// so we can safely use it as a divisor
 				fPosDiff = (float)maxMsgLen / szTmp.length();
-				iPos=iPos*fPosDiff; // cut the string at each cycle
+				iPos=(int) (iPos*fPosDiff); // cut the string at each cycle
 				//printf("OPTIMIZATION: fPosDiff %f, iPos %d\n", fPosDiff, iPos);
 			}
 			//printf("Multi message: %d optimization cyles", iC);
