+++ src/kvirc/ui/kvi_query.cpp
@@ -666,8 +666,8 @@ void KviQuery::ownMessage(const QString &buffer)
 				if(szTmp.length() <= maxMsgLen) break;
 				//if szTmp.length() == 0 we already have break'ed out from here,
 				// so we can safely use it as a divisor
-				fPosDiff = (float)maxMsgLen / szTmp.length();
-				iPos=iPos*fPosDiff; // cut the string at each cycle
+				fPosDiff = (float)maxMsgLen / (float)szTmp.length();
+				iPos=(int) (iPos*fPosDiff); // cut the string at each cycle
 				//printf("OPTIMIZATION: fPosDiff %f, iPos %d\n", fPosDiff, iPos);
 			}
 			//printf("Multi message: %d optimization cyles", iC);
